+++ tests/testthat/test-to_any_case.R
@@ -76,6 +76,11 @@ test_that("complex strings", {
   expect_equal(to_any_case("rStudio", case = "none", prefix = "rrr."),
                "rrr.rStudio")
   
+  expect_equal(to_any_case("Rstudio_STudio_sssTTT", case = "mixed"),
+               "Rstudio_S_Tudio_sss_Ttt")
+  expect_equal(to_any_case("Rstudio_STudio_sssTTT", case = "mixed", parsingoption = 2),
+               "Rstudio_St_udio_sss_Ttt")
+  
 })